@@ -4,20 +4,18 @@
     "lint": "prettier */*.js \"*/**/*{.js,.ts}\" --check"
   },
   "dependencies": {
-    "@project-serum/anchor": "0.25.0",
-    "ts-mocha": "10.0.0"
+    "@project-serum/anchor": "0.25.0"
   },
   "devDependencies": {
-    "@saberhq/anchor-contrib": "1.13.28",
-    "@saberhq/merkle-distributor": "0.3.5",
-    "@saberhq/token-utils": "^1.13.28",
-    "@types/bn.js": "^5.1.0",
-    "@types/chai": "^4.3.0",
-    "@types/mocha": "^9.0.0",
-    "chai": "^4.3.4",
+    "@saberhq/token-utils": "^1.14.9",
+    "@types/bn.js": "^5.1.1",
+    "@types/chai": "^4.3.3",
+    "@types/mocha": "^10.0.0",
+    "chai": "^4.3.6",
     "jsbi": "^4.3.0",
-    "mocha": "^9.0.3",
-    "prettier": "^2.6.2",
-    "typescript": "^4.3.5"
+    "mocha": "^10.1.0",
+    "prettier": "^2.7.1",
+    "ts-mocha": "10.0.0",
+    "typescript": "^4.8.4"
   }
 }
